Working on consolidating the movie scanner code into a single file
