--- server/cmd/api/movies_scanner_helpers.go
@@ -1,42 +1,0 @@
-package main
-
-import (
-	"context"
-	"fmt"
-	"igloo/cmd/internal/database"
-	"igloo/cmd/internal/helpers"
-)
-
-func extractYearFromReleaseDate(releaseDate string) int {
-	if releaseDate == "" {
-		return 0
-	}
-
-	parsed, err := helpers.ParseDate(releaseDate)
-	if err != nil {
-		return 0
-	}
-
-	return parsed.Year()
-}
-
-func (app *Application) getOrCreateArtist(
-	ctx context.Context,
-	qtx *database.Queries,
-	tmdbID int,
-	name string,
-	profilePath string,
-) (*database.Artist, error) {
-	profile := helpers.NullString(profilePath)
-
-	upserted, err := qtx.UpsertArtist(ctx, database.UpsertArtistParams{
-		Name:    name,
-		TmdbID:  int64(tmdbID),
-		Profile: profile,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("upsert artist failed: %w", err)
-	}
-
-	return &upserted, nil
-}
